requirements.txt
requests
sseclient-py
kafka-python
pyspark

pytest